refactor: moves /socks into /inventory
