+++ web-client/src/api/inventory/model.ts
