+++ web-client/src/api/inventory/queries.ts
