+++ web-client/src/api/inventory/service.ts
